Fix the context definition for 32bit linux. (#71)
* Fix the context definition for 32bit linux.

i64 rather than i32 was accidentally used for the general registers.

* Fix mac lints

* Fix deny

* Fix last lint

* Fix android

* Or not

* Sigh macs

* Clippy why

* And slowboi Windows

---------

+++ minidumper/src/ipc/mac.rs
@@ -267,7 +267,7 @@ impl UnixStream {
 
     #[inline]
     pub(crate) fn peek(&self, buf: &mut [u8]) -> io::Result<usize> {
-        self.0.recv_with_flags(buf, libc::MSG_PEEK as i32)
+        self.0.recv_with_flags(buf, libc::MSG_PEEK)
     }
 
     #[inline]
